Move all locking code to a separate class
