+++ src/Kiota.Builder/CodeDOM/CodeClass.cs
@@ -31,10 +31,9 @@ public enum CodeClassKind
 /// <summary>
 /// CodeClass represents an instance of a Class to be generated in source code
 /// </summary>
-public class CodeClass : ProprietableBlock<CodeClassKind, ClassDeclaration>, ITypeDefinition, IDiscriminatorInformationHolder, IDeprecableElement, IDisposable
+public class CodeClass : ProprietableBlock<CodeClassKind, ClassDeclaration>, ITypeDefinition, IDiscriminatorInformationHolder, IDeprecableElement
 {
     protected ConcurrentDictionary<string, CodeProperty> PropertiesByWireName { get; private set; } = new(StringComparer.OrdinalIgnoreCase);
-    private readonly CountdownEvent propertiesBuilt = new(1);
     public bool IsErrorDefinition
     {
         get; set;
@@ -65,36 +64,6 @@ public class CodeClass : ProprietableBlock<CodeClassKind, ClassDeclaration>, ITy
                 AddRange(value);
         }
         get => InnerChildElements.Values.OfType<CodeIndexer>().FirstOrDefault();
-    }
-    public Boolean IsPropertiesBuilt()
-    {
-        return Kind != CodeClassKind.Model || propertiesBuilt.IsSet;
-    }
-    private void WaitForPropertiesBuilt()
-    {
-        if (Kind == CodeClassKind.Model && !Monitor.IsEntered(propertiesBuilt))
-        {
-            propertiesBuilt.Wait();
-        }
-    }
-    public void StartBuildingProperties()
-    {
-        if (Kind == CodeClassKind.Model)
-            Monitor.Enter(propertiesBuilt);
-    }
-    public void PropertiesBuildingDone()
-    {
-        if (Kind == CodeClassKind.Model)
-        {
-            if (!IsPropertiesBuilt())
-            {
-                if (!propertiesBuilt.Signal())
-                {
-                    throw new InvalidOperationException("PropertiesBuilt CountdownEvent is expected to always reach 0 at this point.");
-                }
-            }
-            Monitor.Exit(propertiesBuilt);
-        }
     }
     public override IEnumerable<CodeProperty> AddProperty(params CodeProperty[] properties)
     {
@@ -171,7 +140,6 @@ public class CodeClass : ProprietableBlock<CodeClassKind, ClassDeclaration>, ITy
     {
         ArgumentException.ThrowIfNullOrEmpty(propertyName);
 
-        WaitForPropertiesBuilt();
         if (FindChildByName<CodeProperty>(propertyName, findInChildElements: false) is CodeProperty result)
         {
             return result;
@@ -186,7 +154,6 @@ public class CodeClass : ProprietableBlock<CodeClassKind, ClassDeclaration>, ITy
     {
         ArgumentException.ThrowIfNullOrEmpty(serializationName);
 
-        WaitForPropertiesBuilt();
         if (BaseClass is CodeClass currentParentClass)
             if (currentParentClass.FindPropertyByWireName(serializationName) is CodeProperty currentProperty && !currentProperty.ExistsInBaseType)
                 return currentProperty;
@@ -271,28 +238,9 @@ public class CodeClass : ProprietableBlock<CodeClassKind, ClassDeclaration>, ITy
             _discriminatorInformation = value;
         }
     }
-
     public DeprecationInformation? Deprecation
     {
         get; set;
-    }
-    private bool isDisposed;
-    public void Dispose()
-    {
-        Dispose(true);
-        GC.SuppressFinalize(this);
-    }
-    protected virtual void Dispose(bool disposing)
-    {
-        if (isDisposed) return;
-
-        if (disposing)
-        {
-            // free managed resources
-            propertiesBuilt.Dispose();
-        }
-
-        isDisposed = true;
     }
 }
 public class ClassDeclaration : ProprietableBlockDeclaration
